Add Delta Time measurement

+++ GameLamp/src/Core/Application.h
@@ -12,12 +12,12 @@ namespace Lamp {
 		int run(int argc = 0, char** argv = nullptr);
 
 		virtual void init(int arhc, char** argv) {}
-		virtual void tick(float delta = 0.0f) {}
+		virtual void tick(double delta = 0.0) {}
 		virtual void shutdown() {}
 
 	private:
 
-		void tickInternal(float delta);
+		void tickInternal(double delta);
 
 	private:
 		Window* m_Window = nullptr;
